+++ internal/x/radixtrie/trie_benchmark_test.go
@@ -14,7 +14,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package radixtree
+package radixtrie
 
 import (
 	"testing"
